[#2] Refactor : useGetProduct 로직 수정 및 Skeleton UI 컴포넌트 생성
