+++ .eslintrc.json
@@ -28,6 +28,7 @@
 	"rules": {
 		"no-alert": "off",
 		"react/prop-types": 0,
+		"react/require-default-props": 0,
 		"import/order": [
 			"error",
 			{
